requirements.txt
datasets
transformers
torchtune
torchao

# Single T4 GPU - no Megatron needed
# tensorstore
# zarr